feat: make Text.Iterator unbound
This patch changes iterator semantics: it is now possible to
have `iterator.offset` that doesn't belong to the [from, to) scope.

This makes it convenient when porting code that uses String and Index
onto text iterators.

As a rule of thumb, every iterator advancement method now
non-conditionally mutates iterator. Notable API changes:
- BEFORE: `it.next()` would return `false` when reaching end of the
  range. NOW: `it.next()` doesn't return boolean and always succeeds.
- BEFORE: `it.find()` would return `false` and keep iterator unchanged
  if there's no match. NOW: `it.find()` returns `false`, but moves
  iterator towards the end of the range.

Changes in `it.find()` are convenient to avoid double-traversal of input.