test(jslexer): add 700 esprima tests to ensure tokenization
This patch pulls in esprima fixtures to ensure correct tokenization.
Test expectations are generated with original acron tokenizer.

NOTE: for those tests that make original acorn tokenizer fail,
the expectations are generated with the current version of recovering
lexer.

NOTE-2: it also turned out that support for "blockComment" and "lineComment"
tokens in lexer was added poorly; the patch fixes it.