feature(jslexer): support fast recovery for multiple tokens
Currently, when a new iterator is set to the parser that has reached
EOF, the last token is re-read from the beginning.

This is inefficient in case of long tokens: re-reading will be
O(token size). If amount of `setIterator` calls
is proportional to the length of the token, then it will result in
O(N^2) time.

This patch implements O(1) recovery for the 4 token types
that are usually very long:
- block comments
- line comments
- strings
- templates