feature(wordwrap): speed up word wrap ~5x times
We stop testing every single character, and instead split the string.
Now word wrap performance is much better.

Unfortunately, it does heavily depend on the line width:
the narrow the line, the more lines overall we produce and the more
nodes we add to the tree. It is unclear how to deal with that.