refactor: use editor's highlighter to decorate text
Currently, highlighters add themselves as decoration callbacks to
the editor. This makes it hard for the editor to ensure that there's
only one active highlighter.

This patch inverses the relationship: instead of highlighters
adding themselve as decoration callbacks, editor now uses the active
highlighter to highlight frame.

As a bonus, this gets rid of DefaultHighlighter: when there's no
active highlighter, editor just returns a single "syntax.default" token
for every frame.