chore: stop using Node.JS --experimental-modules flag
A lot of test coverage for Mezzo is driven by Node.js which,
unfortunately, doesn't support ESM out-of-the-box.

There's an `--experimental-modules` node.js flag that provides some
playground, but there's yet to be a decision on how to treat ESM in
Node. Currently, `--experimental-modules` strictly requires `*.mjs`
files.

On the other hand, TypeScript disregards `*.mjs` files altogether - see
https://github.com/Microsoft/TypeScript/issues/27957. As a result, we
can't typecheck our codebase.

In order to do typechecking, we should:
1. switch from `--experimental-modules` onto
[esm](https://github.com/standard-things/esm). This is the current
de-facto approach to using ESM in Node.
2. rename all `.mjs` into `.js`
3. gradually enable `tsc` for Mezzo subfolders.

This patch does Step-1. Note: ESM was allergic to the `__dirname`
variable, so it got renamed into `__DIRNAME`.